Structures | inlineReply fix
