+++ test.js
@@ -1,6 +1,7 @@
 const Discord = require("discord.js");
 const client = new Discord.Client();
 const { GCommands } = require("./src/index");
+const axios = require("axios")
 
 client.on("ready", () => {
     new GCommands(client, {
